requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/dev.txt requirements/dev.in
#
-e git+https://github.com/edx/i18n-tools.git@v0.3.2#egg=i18n_tools==0.3.2
args==0.1.0               # via clint
astroid==1.4.8            # via pylint, pylint-celery, pylint-plugin-utils
click==6.6                # via pip-tools
clint==0.5.1              # via twine
colorama==0.3.7           # via pylint
ddt==1.1.0
edx-lint==0.5.1
first==2.0.1              # via pip-tools
lazy-object-proxy==1.2.2  # via astroid
path.py==8.2.1
pip-tools==1.7.0
pkginfo==1.3.2            # via twine
pluggy==0.3.1             # via tox
polib==1.0.7
py==1.4.31                # via tox
pylint-celery==0.3        # via edx-lint
pylint-django==0.7.2      # via edx-lint
pylint-plugin-utils==0.2.4  # via pylint-celery, pylint-django
pylint==1.5.4             # via edx-lint, pylint-celery, pylint-django, pylint-plugin-utils
pytz==2016.6.1
pyYaml==3.11
requests-toolbelt==0.7.0  # via twine
requests==2.11.0          # via requests-toolbelt, twine
six==1.10.0               # via astroid, edx-lint, pip-tools, pylint
tox-battery==0.2
tox==2.3.1
twine==1.8.1
virtualenv==15.0.3        # via tox
wheel==0.29.0
wrapt==1.10.8             # via astroid

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via twine
